requirements.txt
setuptools~=49.6.0
torchvision~=0.2.2
pillow~=8.1.2
numba~=0.53.1
numpy~=1.20.2
matplotlib~=3.4.1
musicbrainzngs~=0.7.1